server: move TestDecommissionNodeStatus into server_test.go
The test `TestDecommissionNodeStatus` was initially placed in a separate
file `decommission_test.go`, even though it belonged in
`server_test.go`. This was done to avoid import cycles between the
`testcluster` package and the `server` package, as the
`decommission_test.go` file could use the `server_test` package instead.

This commit creates the cluster with `serverutils.StartNewTestCluster()`
instead of `testcluster.StartTestCluster()`, which avoids the import
cycle and allows it to be moved into `server_test.go`. This required
exposing `Server.Decommission()` via `serverutils.TestServerInterface`
as well.

Release justification: non-production code changes
Release note: None
